+++ src/main/java/pl/cantabo/database/profile/ProfileUpdateDTO.java
@@ -24,11 +24,11 @@ public class ProfileUpdateDTO {
 
     private String fontStyle;
 
-    private float margin;
+    private double margin;
 
-    private float maxFont;
+    private double maxFont;
 
-    private float maxMin;
+    private double maxMin;
 
     private String align;
